+++ moshi-ksp/moshi-ksp/src/main/kotlin/dev/zacsweers/moshix/ksp/KspUtil.kt
@@ -1,13 +1,16 @@
 package dev.zacsweers.moshix.ksp
 
+import com.google.devtools.ksp.getAllSuperTypes
 import com.google.devtools.ksp.processing.KSPLogger
 import com.google.devtools.ksp.processing.Resolver
+import com.google.devtools.ksp.symbol.ClassKind.CLASS
 import com.google.devtools.ksp.symbol.KSAnnotated
 import com.google.devtools.ksp.symbol.KSAnnotation
 import com.google.devtools.ksp.symbol.KSClassDeclaration
 import com.google.devtools.ksp.symbol.KSName
 import com.google.devtools.ksp.symbol.KSNode
 import com.google.devtools.ksp.symbol.KSType
+import com.google.devtools.ksp.symbol.Origin.KOTLIN
 import com.google.devtools.ksp.symbol.Visibility
 import com.google.devtools.ksp.symbol.Visibility.INTERNAL
 import com.google.devtools.ksp.symbol.Visibility.JAVA_PACKAGE
@@ -30,6 +33,18 @@ internal fun Resolver.getClassDeclarationByName(fqcn: String): KSClassDeclaratio
 }
 
 internal fun KSClassDeclaration.asType() = asType(emptyList())
+
+internal fun KSClassDeclaration.superclass(resolver: Resolver): KSType {
+  return getAllSuperTypes().firstOrNull {
+    val decl = it.declaration
+    decl is KSClassDeclaration && decl.classKind == CLASS
+  } ?: resolver.builtIns.anyType
+}
+
+internal fun KSClassDeclaration.isKotlinClass(resolver: Resolver): Boolean {
+  return origin == KOTLIN ||
+    hasAnnotation(resolver.getClassDeclarationByName<Metadata>().asType())
+}
 
 internal fun KSAnnotated.hasAnnotation(target: KSType): Boolean {
   return findAnnotationWithType(target) != null
@@ -90,7 +105,7 @@ internal fun KSAnnotation.toAnnotationSpec(resolver: Resolver): AnnotationSpec {
     member.add("%L = ", name)
     when (val value = argument.value!!) {
       resolver.builtIns.arrayType -> {
-//        TODO("Arrays aren't supported tet")
+//        TODO("Arrays aren't supported yet")
 //        member.add("[⇥⇥")
 //        values.forEachIndexed { index, value ->
 //          if (index > 0) member.add(", ")