+++ moshi-sealed/codegen-ksp/src/main/kotlin/dev/zacsweers/moshix/sealed/codegen/ksp/MoshiSealedSymbolProcessor.kt
@@ -10,10 +10,15 @@ import com.google.devtools.ksp.symbol.KSClassDeclaration
 import com.google.devtools.ksp.symbol.KSDeclarationContainer
 import com.google.devtools.ksp.symbol.Modifier
 import com.squareup.kotlinpoet.AnnotationSpec
+import com.squareup.kotlinpoet.CodeBlock
+import com.squareup.kotlinpoet.FileSpec
+import com.squareup.kotlinpoet.MemberName
+import com.squareup.kotlinpoet.asClassName
 import com.squareup.moshi.JsonClass
 import dev.zacsweers.moshix.sealed.annotations.DefaultNull
 import dev.zacsweers.moshix.sealed.annotations.DefaultObject
 import dev.zacsweers.moshix.sealed.annotations.TypeLabel
+import dev.zacsweers.moshix.sealed.runtime.internal.ObjectJsonAdapter
 
 @AutoService(SymbolProcessor::class)
 public class MoshiSealedSymbolProcessor : SymbolProcessor {
@@ -106,50 +111,63 @@ public class MoshiSealedSymbolProcessor : SymbolProcessor {
     val defaultObjectAnnotation = resolver.getClassDeclarationByName<DefaultObject>().asType()
     val typeLabelAnnotation = resolver.getClassDeclarationByName<TypeLabel>().asType()
     val useDefaultNull = type.hasAnnotation(defaultNullAnnotation)
+    val objectAdapters = mutableListOf<CodeBlock>()
     val sealedSubtypes = type.sealedSubtypes()
-      .mapNotNullTo(LinkedHashSet()) { subtype ->
+      .mapTo(LinkedHashSet()) { subtype ->
         val className = subtype.toClassName()
-        if (subtype.classKind == OBJECT) {
-          if (subtype.hasAnnotation(defaultObjectAnnotation)) {
-            if (useDefaultNull) {
-              // Print both for reference
-              error("""
-                  Cannot have both @DefaultNull and @DefaultObject. @DefaultObject type: $type
-                  Cannot have both @DefaultNull and @DefaultObject. @DefaultNull type: $subtype
-                """.trimIndent())
-            } else {
-              return@mapNotNullTo Subtype(
-                className = className,
-                isDefaultObject = true,
-                typeLabelType = null,
-                originatingElement = null
-              )
-            }
-          } else if (!useDefaultNull) {
-            error("Unhandled object type, cannot serialize this. $type")
+        val isObject = subtype.classKind == OBJECT
+        if (isObject && subtype.hasAnnotation(defaultObjectAnnotation)) {
+          if (useDefaultNull) {
+            // Print both for reference
+            error("""
+                Cannot have both @DefaultNull and @DefaultObject. @DefaultObject type: $type
+                Cannot have both @DefaultNull and @DefaultObject. @DefaultNull type: $subtype
+              """.trimIndent())
+          } else {
+            return@mapTo Subtype.ObjectType(className)
           }
-          return@mapNotNullTo null
         } else {
           val labelAnnotation = subtype.findAnnotationWithType(typeLabelAnnotation)
             ?: error("Missing @TypeLabel: $subtype")
-          val labelValue = labelAnnotation.arguments
-            .find { it.name?.getShortName() == "value" }
+
+          val labels = mutableListOf<String>()
+
+          labels += labelAnnotation.arguments
+            .find { it.name?.getShortName() == "label" }
             ?.value as? String
-            ?: error("No value found for label annotation!")
-          Subtype(className, false, labelValue, null)
+            ?: error("No label member for TypeLabel annotation!")
+
+          @Suppress("UNCHECKED_CAST")
+          labels += labelAnnotation.arguments
+            .find { it.name?.getShortName() == "alternateLabels" }
+            ?.value as? List<String> // arrays are lists in KSP https://github.com/google/ksp/issues/135
+            ?: emptyList() // ksp ignores undefined args https://github.com/google/ksp/issues/134
+
+          if (isObject) {
+            objectAdapters.add(CodeBlock.of(
+              ".%1M<%2T>(%3T(%2T))",
+              MemberName("com.squareup.moshi", "addAdapter"),
+              className,
+              ObjectJsonAdapter::class.asClassName()
+            ))
+          }
+          Subtype.ClassType(className, labels)
         }
       }
 
-    createType(
+    val preparedAdapter = createType(
       targetType = type.toClassName(),
       isInternal = Modifier.INTERNAL in type.modifiers,
       typeLabel = typeLabel,
       useDefaultNull = useDefaultNull,
       originatingElement = null,
       generatedAnnotation = generatedAnnotation,
-      subtypes = sealedSubtypes
+      subtypes = sealedSubtypes,
+      objectAdapters = objectAdapters
     )
-      .writeTo(codeGenerator)
+
+    preparedAdapter.spec.writeTo(codeGenerator)
+    preparedAdapter.proguardConfig.writeTo(codeGenerator)
   }
 
   private fun KSClassDeclaration.sealedSubtypes(): Set<KSClassDeclaration> {
@@ -175,3 +193,5 @@ public class MoshiSealedSymbolProcessor : SymbolProcessor {
 
   }
 }
+
+internal data class PreparedAdapter(val spec: FileSpec, val proguardConfig: ProguardConfig)