+++ moshi-sealed/runtime/src/main/kotlin/dev/zacsweers/moshix/sealed/annotations/TypeLabel.kt
@@ -1,0 +1,30 @@
+/*
+ * Copyright (c) 2020 Zac Sweers
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package dev.zacsweers.moshix.sealed.annotations
+
+import kotlin.annotation.AnnotationRetention.RUNTIME
+import kotlin.annotation.AnnotationTarget.CLASS
+
+/**
+ * Use this annotation to specify the type label of a subtype.
+ *
+ * @property label the type label value.
+ * @property alternateLabels an optional array of other values this type could be represented by.
+ */
+@Target(CLASS)
+@Retention(RUNTIME)
+public annotation class TypeLabel(val label: String, val alternateLabels: Array<String> = [])